feat: StateMachine architecture refactor; StateMachine deleted; PlayerStateMachine deleted; StateMachineController created; UnitState refactor;
