+++ Assets/Scenes/SampleScene.unity
@@ -9211,7 +9211,6 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 9b54f9545555f744c8c9665d7cde94b7, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  player: {fileID: 0}
   inventorySystem: {fileID: 1026977245}
 --- !u!4 &1026977244
 Transform:
@@ -9260,6 +9259,7 @@ GameObject:
   - component: {fileID: 1139493140}
   - component: {fileID: 1139493141}
   - component: {fileID: 1139493142}
+  - component: {fileID: 1139493139}
   m_Layer: 8
   m_Name: Player
   m_TagString: Player
@@ -9382,6 +9382,18 @@ SpriteRenderer:
   m_WasSpriteAssigned: 1
   m_MaskInteraction: 0
   m_SpriteSortPoint: 0
+--- !u!114 &1139493139
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1139493134}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: fd9d1c75b4f541b469dab8f28d71121d, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!95 &1139493140
 Animator:
   serializedVersion: 3
@@ -9413,6 +9425,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 0adc9d93edfbb2144b7401ffbe4ce3cd, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  playerStateMachine: {fileID: 1139493139}
 --- !u!114 &1139493142
 MonoBehaviour:
   m_ObjectHideFlags: 0
