--- Assets/Scripts/UnityPowerCore/Objects/StateMachine/StateMachine.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 06b9861f0842a56479a7ca5d40471cd8
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
