+++ Assets/Scripts/UnityPowerCore/Objects/StateMachine/StateMachineController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: dd8ffa96e6e32b14aa62c7f641206528
+guid: e2de6cbe0c6ed9e41b7a095a47818946
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
