--- Assets/Scripts/UnityPowerCore/Objects/StateMachine/StateMachineObject.cs
@@ -1,6 +1,0 @@
-﻿public interface StateMachineObject
-{
-    void objectStart();
-    void objectUpdate();
-    void objectFixedUpdate();
-}
