--- Assets/Scripts/UnityPowerCore/Objects/StateMachine/StateMachineObject.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4a13d9d762ea27043bb9e74bebaf4217
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
